Added more tests.

+++ src/myun/type/PrimitiveTypes.java
@@ -11,7 +11,8 @@ public final class PrimitiveTypes {
     public static final String MYUN_INT = "Int";
     public static final String MYUN_FLOAT = "Float";
 
+    // we assume 64-bit architecture
     public static final String LLVM_BOOL = "i1";
-    public static final String LLVM_INT = "i32";
-    public static final String LLVM_FLOAT = "float";
+    public static final String LLVM_INT = "i64";
+    public static final String LLVM_FLOAT = "double";
 }
